app_OCR_doctr/requirements.txt
fastapi
uvicorn
python-doctr
python-multipart